requirements.txt
swarms>=5.0.0
pillow>=10.0.0
diffusers>=0.25.0
transformers>=4.36.0
torch>=2.0.0
controlnet-aux>=0.0.7
accelerate>=0.25.0
opencv-python>=4.8.0
numpy>=1.24.0
cairosvg>=2.7.0
svgwrite>=1.4.3
openai>=1.0.0
python-dotenv>=1.0.0
